Add primitive filtering function to Stream.
It should be possible to define filters on osmium.Stream so that objects not
matching the filter don't have to be given to Javascript at all. This is
quite an important optimization. This code only  has a simplified version
of such a filter and is itself quite expensive, because the filter has to
be given from Javascript to C++ on each next() call.