Always reuse the same wrapper object.
The Node/Way/Relation/Area/Changeset wrapper objects are only created once at
program start and are re-used again and again. This only works for the
streaming interface, not the (deprecated) handler interface. We can do that
because those objects are usable only inside the callback anyway.
This change improves performance by about 20%.