Fix issue with data not being flushed.
Node-osmium buffers data in some places, and those buffers weren't flushed
properly in every case. Problem is we can not flush them automatically in all
places, because we do not know when to do this. So now the user must call
`end()` on any handlers he uses to flush everything.

See the demo/multipolygon/index.js for an example on how this is done.

This change might break some existing code where the flush was done
automatically!

See #60.